+++ src/main/java/ladder/utils/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.utils;
 
 public interface BooleanGenerator {
