+++ src/test/java/ladder/domain/LadderTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.utils.BooleanGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
